Exclude file with defines instead of csproj
